Fix bug 72475 (Binlog events with binlog_format=MIXED are unconditionally logged in ROW format)
The bug is a result of an old fix for bug 20499. This fix changes
mixed replication behavior as follows:

If using MIXED mode replication, all DDL and DML statements that touch
a temporary table will be marked to binlog in ROW format. This means
that any statement that is using exclusively temporary tables will not
be binlogged at all. Any DDL or DML that uses a mix of temporary and
non-temporary tables will be binlogged in ROW format for the
non-temporary table operations. The one exception is DROP table when
dropping a temporary table which will always be binlogged as DROP
TEMPORARY TABLE IF EXISTS on the master.