Bug 1455432: Audit log plugin should allow to include or exclude logging for specific users
Added two global variables:
- audit_log_include_accounts: accounts to include in audit logging.
- audit_log_exclude_accounts: accounts to exclude from audit logging.

The value can be NULL, empty string or comma separated list of
accounts in form user@host or 'user'@'host' (if user or host contains
comma).

Only one variable can be not empty at a time. If one is set to be not
empty, another one becomes empty.